+++ src/constants/components.ts
@@ -1,0 +1,4 @@
+export const enum ChatBubbleEdgeType {
+  LEFT = 'LEFT',
+  RIGHT = 'RIGHT',
+}
