+++ src/index.tsx
@@ -4,12 +4,14 @@ import { createRoot } from 'react-dom/client';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import { RecoilRoot } from 'recoil';
 import NavBar from './components/Navigation/NavBar';
-import Step from './components/SignIn/Step';
+import Gender from './components/SignUp/Gender';
+import Location from './components/SignUp/Location';
+import Name from './components/SignUp/Name';
 import ChatRoom from './pages/ChatRoom';
 import ChatRoomDescription from './pages/ChatRoomDescription';
 import Main from './pages/Main';
 import MyPage from './pages/MyPage';
-import SignIn from './pages/SignIn';
+import SignUp from './pages/SignUp';
 import baseTheme from './styles/baseTheme';
 import GlobalStyles from './styles/globalStyles';
 import Playground from './pages/Playground';
@@ -31,9 +33,10 @@ function RootWithCallbackAfterRender() {
               <Route path="playground" element={<Playground />} />
               <Route path="chatroom" element={<ChatRoom />} />
               <Route path="starting-chatroom" element={<ChatRoomDescription />} />
-              <Route path="sign-in/*" element={<SignIn />}>
-                <Route path=":step" element={<Step />} />
-                <Route path="*" element={<Step props={1} />} />
+              <Route path="signup" element={<SignUp />}>
+                <Route path="name" element={<Name />} />
+                <Route path="gender" element={<Gender />} />
+                <Route path="location" element={<Location />} />
               </Route>
               <Route path="mypage" element={<MyPage />} />
             </Routes>
@@ -43,6 +46,7 @@ function RootWithCallbackAfterRender() {
     </React.Fragment>
   );
 }
+
 const container = document.getElementById('root');
 const root = createRoot(container);
 root.render(<RootWithCallbackAfterRender />);
