+++ src/utils/index.ts
@@ -1,3 +1,4 @@
-export * from './getData';
 export * from './debounce';
+export * from './getData';
 export * from './throttle';
+export * from './verifyInput';
