+++ src/utils/verifyInput.ts
@@ -1,0 +1,3 @@
+export const verifyInput = (isRegExp: RegExp, value: string) => {
+  return isRegExp.test(value);
+};
